Added check for invalid issue and volume keys

+++ controllers/archiveController.js
@@ -2,6 +2,8 @@ const Archive = require("../models/Archive");
 
 const addNewArchive = async (req, res) => {
   const { volume, issue, file } = req.body;
+  if (isNaN(volume) || isNaN(issue))
+    return res.status(400).json({ error: "Invalid volume or issue" });
   const alreadyExistingArchive = await Archive.find({ volume, issue });
   if (alreadyExistingArchive) await alreadyExistingArchive.deleteOne();
   await Archive.create({ volume, issue, file });
